smart-capex-capacity/src/step_09_capacity_economical_pipeline/requirements.txt
prophet==1.1
holidays==0.14.2
google-cloud-aiplatform==1.59.0
kfp==2.5.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-storage==2.17.0
pandas==2.2.2
numpy==1.26.2
numpy-financial==1.0.0
db-dtypes==1.2.0
fsspec==2024.6.1
gcsfs==2024.6.1
pandas-gbq==0.23.1
pyarrow==16.1.0
openpyxl==3.1.5
joblib==1.4.2
scipy==1.14.0
multiprocess==0.70.16
haversine==2.8.1
scikit-learn==1.5.1
xgboost==2.1.0